created Model; refactored components and sources; removed parameter fixing
