+++ scarlet/component.py
@@ -3,105 +3,25 @@ from functools import partial
 import autograd.numpy as np
 
 from .frame import Frame
+from .model import Model
 from .parameter import Parameter
 from .bbox import Box, overlapped_slices
 
 
-def _get_parameter(cls, name, *parameters):
-    # check parameters first during optimization
-    for p in parameters:
-        # if parameters are autograd ArrayBoxes
-        # need to access the wrapped class with _value
-        if p._value.name == name:
-            return p
-
-    # find them from self (use all even if fixed!)
-    names = tuple(p.name for p in cls._parameters)
-    try:
-        return cls._parameters[names.index(name)]
-    except ValueError:
-        return None
-
-
-class Component(ABC):
-    """A single component in a blend.
-
-    This class acts as base for building a complex :class:`scarlet.blend.Blend`.
-
-    Parameters
-    ----------
-    model_frame: `~scarlet.Frame`
-        The spectral and spatial characteristics of the model.
-    parameters: list of `~scarlet.Parameter`
-    bbox: `~scarlet.Box`
-        Hyper-spectral bounding box of this component
-    kwargs: dict
-        Auxiliary information attached to this component.
-    """
-
-    def __init__(self, model_frame, *parameters, bbox=None, **kwargs):
-
+class Component(Model):
+    def __init__(self, frame, parameters=None, children=None, bbox=None, **kwargs):
         if bbox is None:
-            bbox = model_frame.bbox
-        self._bbox = bbox
-        self.set_model_frame(model_frame)
+            bbox = frame.bbox
 
-        if hasattr(parameters, "__iter__"):
-            for p in parameters:
-                assert isinstance(p, Parameter)
-            self._parameters = parameters
-        else:
-            assert isinstance(parameters, Parameter)
-            self._parameters = tuple(parameters)
-        self.check_parameters()
+        # component should always be a data cube
+        assert bbox.D == frame.bbox.D
 
-        # additional non-optimization parameters of component
-        self.kwargs = kwargs
+        super().__init__(
+            frame, parameters=parameters, children=children, bbox=bbox, **kwargs
+        )
+        self.set_model_frame(frame)
 
-    @property
-    def bbox(self):
-        """Hyper-spectral bounding box of this component (Channel, Height, Width)
-        """
-        return self._bbox
-
-    @property
-    def parameters(self):
-        """The list of non-fixed parameters
-
-        Returns
-        -------
-        list of parameters available for optimization
-        If `parameter.fixed == True`, the parameter will not returned here.
-        """
-        return tuple(p for p in self._parameters if not p.fixed)
-
-    @property
-    def parameter_names(self):
-        names = tuple(p.name for p in self._parameters if not p.fixed)
-
-    def get_parameter(self, name, *parameters):
-        return _get_parameter(self, name, *parameters)
-
-    @abstractmethod
-    def get_model(self, *parameters, frame=None):
-        """Get the model for this component
-
-        Parameters
-        ----------
-        parameters: tuple of optimimzation parameters
-
-        frame: `~scarlet.frame.Frame`
-            Frame to project the model into. If `frame` is `None`
-            then the model contained in `bbox` is returned.
-
-        Returns
-        -------
-        model: array
-            (Channels, Height, Width) image of the model
-        """
-        pass
-
-    def set_model_frame(self, model_frame):
+    def set_model_frame(self, frame):
         """Sets the frame for this component.
 
         Each component needs to know the properties of the Frame and,
@@ -109,27 +29,12 @@ class Component(ABC):
 
         Parameters
         ----------
-        model_frame: `~scarlet.Frame`
+        frame: `~scarlet.Frame`
             Frame of the model
         """
-        self.model_frame = model_frame
         self.model_frame_slices, self.model_slices = overlapped_slices(
-            model_frame.bbox, self.bbox
+            frame.bbox, self.bbox
         )
-
-    def check_parameters(self):
-        """Check that all parameters have finite elements
-
-        Raises
-        ------
-        `ArithmeticError` when non-finite elements are present
-        """
-        for k, p in enumerate(self._parameters):
-            if not np.isfinite(p).all():
-                msg = "Component {}, Parameter '{}' is not finite:\n{}".format(
-                    self.__class__.__name__, p.name, p
-                )
-                raise ArithmeticError(msg)
 
     def model_to_frame(self, frame=None, model=None):
         """Project a model into a frame
@@ -155,8 +60,8 @@ class Component(ABC):
         if model is None:
             model = self.get_model()
         # Use the full model frame by default
-        if frame is None or frame == self.model_frame:
-            frame = self.model_frame
+        if frame is None or frame == self.frame:
+            frame = self.frame
             frame_slices = self.model_frame_slices
             model_slices = self.model_slices
         else:
@@ -171,27 +76,6 @@ class Component(ABC):
         return result
 
 
-class Factor(ABC):
-    def __init__(self, model_frame, *parameters, bbox=None):
-        self._parameters = parameters
-        self._bbox = bbox
-
-    @property
-    def parameters(self):
-        return self._parameters
-
-    @property
-    def bbox(self):
-        return self._bbox
-
-    @abstractmethod
-    def get_model(self, *parameters):
-        pass
-
-    def get_parameter(self, name, *parameters):
-        return _get_parameter(self, name, *parameters)
-
-
 class FactorizedComponent(Component):
     """A single component in a blend.
 
@@ -199,7 +83,7 @@ class FactorizedComponent(Component):
 
     Parameters
     ----------
-    model_frame: `~scarlet.Frame`
+    frame: `~scarlet.Frame`
         The spectral and spatial characteristics of the full model.
     bbox: `~scarlet.Box`
         Hyper-spectral bounding box of this component.
@@ -209,33 +93,14 @@ class FactorizedComponent(Component):
         Parameterization of the morphology.
     """
 
-    def __init__(self, model_frame, spectrum, morphology):
+    def __init__(self, frame, spectrum, morphology, **kwargs):
+        from .spectrum import Spectrum
+        from .morphology import Morphology
+
         assert isinstance(spectrum, Factor)
         assert isinstance(morphology, Factor)
         bbox = spectrum.bbox @ morphology.bbox
-        parameters = spectrum.parameters + morphology.parameters
-        super().__init__(model_frame, *parameters, bbox=bbox)
-        self._spectrum = spectrum
-        self._morphology = morphology
-
-    @property
-    def spectrum(self):
-        """Numpy view of the component SED
-        """
-        return self._spectrum.get_model()
-
-    @property
-    def morphology(self):
-        """Numpy view of the component morphology
-        """
-        return self._morphology.get_model()
-
-    @property
-    def center(self):
-        if hasattr(self._morphology, "center"):
-            return self._morphology.center
-        else:
-            return None
+        super().__init__(frame, children=[spectrum, morphology], bbox=bbox, **kwargs)
 
     def get_model(self, *parameters, frame=None):
         """Get the model for this component.
@@ -253,61 +118,64 @@ class FactorizedComponent(Component):
         model: array
             (Channels, Height, Width) image of the model
         """
-        spectrum = self._spectrum.get_model(*parameters)
-        morph = self._morphology.get_model(*parameters)
-        model = spectrum[:, None, None] * morph[None, :, :]
+        spectrum, morphology = self.get_models_of_children(*parameters)
+        model = spectrum[:, None, None] * morphology[None, :, :]
         # project the model into frame (if necessary)
         if frame is not None:
             model = self.model_to_frame(frame, model)
         return model
 
 
-class CubeComponent(Component):
-    """A single component in a blend.
+from autograd.extend import defvjp, primitive
 
-    Uses full cube parameterization.
 
-    Parameters
-    ----------
-    frame: `~scarlet.Frame`
-        The spectral and spatial characteristics of this component.
-    cube: `~scarlet.Parameter`
-        3D array (C, Height, Width) of the initial data cube.
-    bbox: `~scarlet.Box`
-        Hyper-spectral bounding box of this component.
+@primitive
+def _add_models(*models, full_model, slices):
+    """Insert the models into the full model
+
+    `slices` is a tuple `(full_model_slice, model_slices)` used
+    to insert a model into the full_model in the region where the
+    two models overlap.
     """
+    for i in range(len(models)):
+        if hasattr(models[i], "_value"):
+            full_model[slices[i][0]] += models[i][slices[i][1]]._value
+        else:
+            full_model[slices[i][0]] += models[i][slices[i][1]]
+    return full_model
 
-    def __init__(self, frame, cube, bbox=None):
-        assert isinstance(cube, Parameter) and cube.name == "cube"
-        parameters = (cube,)
-        super().__init__(frame, *parameters, bbox=bbox)
 
-    @property
-    def cube(self):
-        return self.get_parameter("cube")._data
+def _grad_add_models(upstream_grad, *models, full_model, slices, index):
+    """Gradient for a single model
 
-    def get_model(self, *parameters, frame=None):
-        cube = self.get_parameter("cube", *parameters)
-        if frame is not None:
-            cube = self.model_to_frame(frame, cube)
-        return cube
+    The full model is just the sum of the models,
+    so the gradient is 1 for each model,
+    we just have to slice it appropriately.
+    """
+    model = models[index]
+    full_model_slices = slices[index][0]
+    model_slices = slices[index][1]
+
+    def result(upstream_grad):
+        _result = np.zeros(model.shape, dtype=model.dtype)
+        _result[model_slices] = upstream_grad[full_model_slices]
+        return _result
+
+    return result
 
 
 class CombinedComponent(Component):
     def __init__(self, model_frame, components, mode="add"):
         if hasattr(components, "__iter__"):
-            self.components = components
+            assert all(isinstance(c, Component) for c in components)
         else:
-            self.components = (components,)
+            assert isinstance(components, Component)
+            components = (components,)
 
-        assert mode in ["add", "multiply"]
+        assert mode in ["add"]  # , "multiply"]
         self.mode = mode
 
-        parameters = []
-        for c in self.components:
-            parameters += c.parameters
-
-        super().__init__(model_frame, *parameters, bbox=self.bbox)
+        super().__init__(model_frame, children=components, bbox=self.bbox)
 
     def __getitem__(self, i):
         return self.components.__getitem__(i)
@@ -354,3 +222,35 @@ class CombinedComponent(Component):
         if frame is not None:
             model = self.model_to_frame(frame, model)
         return model
+
+
+class CubeComponent(Component):
+    """A single component in a blend.
+
+    Uses full cube parameterization.
+
+    Parameters
+    ----------
+    frame: `~scarlet.Frame`
+        The spectral and spatial characteristics of this component.
+    cube: `~scarlet.Parameter`
+        3D array (C, Height, Width) of the initial data cube.
+    bbox: `~scarlet.Box`
+        Hyper-spectral bounding box of this component.
+    """
+
+    def __init__(self, frame, cube, bbox=None, **kwargs):
+        if isinstance(image, Parameter):
+            assert cube.name == "cube"
+        else:
+            constraint = PositivityConstraint()
+            cube = Parameter(
+                cube, name="cube", step=relative_step, constraint=constraint
+            )
+        super().__init__(frame, cube, bbox=bbox, **kwargs)
+
+    def get_model(self, *parameters, frame=None):
+        cube = self.get_parameter(0, *parameters)
+        if frame is not None:
+            cube = self.model_to_frame(frame, cube)
+        return cube
